# perf/refactor: small optimizations for the big rewrite