--- a/src/resolver/lib/lib.ts
+++ b/src/resolver/lib/lib.ts
@@ -93,3 +93,7 @@ export function mapToObject<T>(map: Map<string, T>): { [key: string]: T } {
 	}
 	return o
 }
+
+export function compareStrings(a: string, b: string): number {
+	return a > b ? 1 : a < b ? -1 : 0
+}