* **What kind of change does this PR introduce?** (Bug fix, feature, docs update, ...)
a few small changes on top of the rewrite, which according to my measurements (using `yarn unit performance`) improve the performance by a few percent


* **What is the current behavior?** (You can also link to an open issue here)



* **What is the new behavior (if this is a feature change)?**



* **Other information**: