@nytamin 2023-08-14
Thanks @ianshade !
I've tested each commit, and then cherry picked most of them, see below:

* perf: turn debug methods into no-op by default
  _I couldn't see any performance improvement on this, so I discarded this._
* perf: possible improvements in comparisons
  _Improvement: 8.04ms -> 7.60ms, cherry-picked!_
* perf: possibly faster hash building
  _I extracted the changes to the hash-function. Improvement: 925μs -> 840μs, cherry-picked!_
* perf: use slice() instead of spread to clone arrays
  _I couldn't see any performance improvement on this, so I discarded this_
* refactor: replace if ... else if ... chain with switch
  _No performance change, but better readability! Cherry-picked!_
* refactor: extract comparison function
  _Slight performance gain: 267μs -> 258μs, cherry-picked!_
* perf: use regex match only when needed
  _Slight performance gain: 115μs -> 96μs, cherry-picked!_
* refactor: replace regular for loops with for ... of 
  _No performance change, but easier to read, cherry-picked!_
* refactor: less deep property access and regular for loops
  _I couldn't see any performance improvements on this. I think that the "undeep properties" reduces readability somewhat, so I discarded this._